Added content-sensitive settings: status, first interval, regular interval
Added the following options to the widget (and test project):
- Interval Attribute: an optional attribute specifying the timer
interval, overrrides the standard interval if set. Widget responds to
changes of this value
- First interval attribute: an optional attribute to specify a customer
delay for the first tick. Useful for restarting a clock, ticking after 5
seconds, then every 60 seconds thereafter.
- Timer status attribute: an optional attribute specifying whether the
timer is running or not. Widget responds to changes of this attribute.